Correctly reload SSP Viewer with the reload button.
OSCALLoader.js initially sets the variable isLoaded to false, and uses this variable
to resolve the controls for an SSP. This caused a crash as referenced in EGRC-298.
OSCALLoader.js now resets isLoaded to false when the reload button is pressed.
SSP Viewer reload button is now working as intended.