Move top-level control modification icon next to control title
OSCALControlPart.js renders the modification icon for each control part.
Before, this rendered the icon under the title, which was undesirable as referenced in EGRC-306.
OSCALControl.js now renders the icon next to the title, and passes in null as the modifications argument
to its top-level control part so that the icon is not rendered in OSCALControlPart.js
Also, in OSCALControlModification.js the <div> around the icon was removed so that it would render next to the title.