Handle different control modification definitions
FedRAMP & NIST are defining control modifications in slightly different
ways. This on top of even more changes to FedRAMP's definition of a
control modification caused OSCALProfiles to load incorrectly. We now
deal with those changes in a cleaner and less repetitive way. This
should make dealing with changes in the future easier as well.