Display proper information for add/remove alters
The getAlterAddsOrRemovesDisplay() function assumed that add and
remove alters would be displayed in a similar way. The function now
checks wheter the alter adds or removes and deals with it accordingly.
Add alters also treated props and parts similarly. It now displays all
props that are contained within a part object.