Correct Control Implementation tooltip test
Styling was previously included in the title (description) of a tooltip
for consistency, which caused a test to fail. When styling is excluded
in the title of an accessible role, the role's name is based on the
title rather than the label. Otherwise the label is selected as the
role name. The test has been updated to this description.