Document contribution process
We need basic documentation on how to contribute in order to facilitate
a community around this project. This is early documentation which is
likely to grow and improve quite a bit going forward. Initially, we want
to ask that contributors create issues prior to PRs and give some
general guidance on what the process will look like after a PR has been
opened.