Use uuid to generate unique keys.
OSCALControl was using an index as a key for each OSCALControlPart
created. This is advised against by react. The parts being mapped do not
have a clear id or field to use as a unique key. Using uuid is the next
best way of giving each child element of a list a unique key that
persists until the child is recreated.