Update stories for Metadata and System Implementation Components
Changed .stories.js files to incorporate all required
props in the default and then modularly add optional
props with each variation.

Added the display of parties to the Metadata story
since parties are an optional prop.

Altered the System Implementation story to have
a variation of the display of the version, a
property within components.