Organize storybook and test data.
Previously, all of the data was in various files, even duplicated
in some instances (across not within files). Added a new directory,
test-data, which contains the 7 files where all of the data for
testing and the storybook are located.

When creating the files, removed the duplications of the data by
placing the duplications into one variable. Separated the data by
category for ease of finding the data when needed.