Fix linting errors in the PR checks
In the code checks for the pull request, there is a linting
error regarding the package-lock JSON files. It states that
local changes to those files would be overwritten by a
checkout and the changes must be committed. The package-lock
JSON files were updated and committed to resolve the linting
errors.