Clean Up Storybook Docs
Combined the two Control Prose components
into one component with multiple stories.

Changed the Control story to display a
top-level modification instead of a control part modification.

Renamed "Control Display" to "Control".

Changed the name of some stories to match the format
"With ..." which is more consistent and readable.

Changed filename of OSCALModifications.stories.js to
OSCALControlModifications.stories.js.