Change BackMatter and Metadata Storybook Docs
Altered BackMatter.stories variations. Now, the
default variation starts with a title and more
props are added with each variation. The last
variation is without the display of a title.

Changed the name of the stories to match the format
"With..." which is more consistent and readable.

Combined Metadata.stories variations into one. Now,
the variation displays the parties and roles.