Update BackMatter, Metadata, and System Implementation Stories
Removed the name of the components in front
of "With" in the variations to. lign with
the rest of the documentation. For example,
"BackMatterWithDescription" is now
"WithDescription". Metadata.stories were
changed also based on this.

Consolidated the SystemImplementation variation
into the default. Now, the default displays
the version and the last modified values.

Changed SystemData.js based on consolidating
the SystemImplementation.stories.