Clean Up Storybook Docs and Naming
Changed the titles of Storybook stories for Control Part
and Control Implentation stories to be consistent
with the rest of the stories.

Changed test data for Modifications to display slightly more
verbose information on Storybook.

Edited the README to flow better by moving the link to
Storybook further up into the Intro section.

Deleted the unused file
stories/OSCALControlProseByComponent.stories.js