Clean Up Control Prose
Removed a console error caused by a `title` prop
being undefined under certain circumstances in
`OSCALReplacedProseWithByComponentParameterValue`
this was done by a quick workaround adding a conditional
if the prop is undefined.

Also added the `componentId` prop into the ImplReq test data
to work with this change and better document the props
for Control Prose.