Update statement fetch based on FedRAMP SSP
The FedRAMP SSP template provides a different format to how statement
set-parameters are provide than seen in NIST's SSPs. Instead of having
set-parameters provided on each statement of a control, they are
provided by only the initial statement of a control in this SSP
template. Control and underlying components have been reworked to fit
this description.