Fix Modification Rendering Issues
Change the way the `isLoaded` state variable is set to
fix an issue where modifications would not render in the
Component Viewer.
`isLoaded` was being set to true upon the modifications for the
first Control Implementation being resolved, which was causing
the state to be "loaded" when not all the modifications were loaded.
Added a `pendingProcesses` counter to track when all of the
implementations have had their modifications resolved before
setting the isLoaded state.