Add Awaits to Tests to Correctly Test with the New Skeleton
The Back Matter and Metadata components previously were rendered
alonside the skeleton. Now that those components do not render right
away, the tests for those components had to be updated; the
test functions were changed to use `await` and a timeout of 10000ms
to wait on the page to load before finding the elements.