Modify Profile Test to Fix Intermittent Failures
Changed the profile test to use waitFor() instead of an await
on findByText().
Occasionally, certain conditions caused findByText() to throw an error,
likely due to the long load times for the imported controls in the
profile viewer. The errors caused the tests to fail even though the
code should pass.
This change should resolve these errors as waitFor() retries getByText()
until it does not throw an error, more resiliently handling the
conditions that caused findByText() to throw errors.