Change Timeouts on Profile Test
Change the timeouts on the profile test to explicity be 5000ms.
Before, the explicit timeout was longer than the default test
timeout of 5000ms set by jest. This was causing a less descriptive
error message to appear on a failed test.
This change sets the timeouts to 5000ms and sets the overall test
timeout to the slightly higher 6000ms in order to keep the optimal
timeout time of 5000ms while also producing the correct error message
on a failed timeout.