Check for Media Type in Profile Resolver
Modify the profile resolver to check for the `media-type` value
when grabbing a resource from back matter. This is a fix for
an issue where the first rlink in each resource was found even if there
were multiple links. This potentially could have lead to the wrong link
being grabbed and thus failing to load the resource.