Add Test for Profile Resolver
Added a test to `OSCALProfile.test.js` to test if the `imports`
of the test profile are valid by checking if `getUriFromBackMatterByHref`
returns a string.
This test is added to verify that the fix to the way resources are grabbed
by the profile resolver is not broken by future changes.