Modify Test Data to Test for Multiple Links in Back Matter
Changed the test data to include multiple rlinks in order to properly
test that all features work with the new fix.
Made a small fix to the back matter component to give each rlink
a unique key when there are more than one.

Also removed the unnecessary test for the profile resolver, as that
functionality is tested within the existing profile tests.