Modify names and add manual trigger
Modify commit messages when updating the library
and viewer for better specification of what was
done.

Update commit email and name to the organization
email and name because the updates are made by
the organization's workflow file in this
repository and not directly by the person who
triggered the workflow.