Delete and recreate packages on push
Delete and recreate the oscal-react-library package on a push
to the repository, negating the need to update package.json
and package-lock.json files and make a second push prior to
publishing the package. Update the package.json and
package-lock.json files to an earlier version number as the
version numbers after 0.1.0 are no longer being used.