Consolidate GitHub Packages
Move the production build of the Oscal Viewer, previously contained in a
separate `oscal-viewer` package, into the `oscal-react-library` package,
and remove the viewer package.

This fixes issues with the oscal-viewer package, which was previously
set up only for offline development by using local paths in its dependencies,
which are not compatible with publishing to an online registry.
Moving everything into one package allows for the end user to install both the
Viewer and its dependencies all in one place.