Remove Production Build from OSCAL Library Package
Change the "files" in the OSCAL React Library package.json to
no longer include the production build of the Viewer, so that the
publicly available package will not force the Viewer along with the
Library.

Also change the GitHub actions workflow to accomodate this change by
no longer building the production build.