Limit Inputs to Feature or Bug Fix Releases
Replace the "version" input with a "type" input to narrow down the
functionality of the workflow. This makes it so that only `npm version patch`
and `npm version minor` can properly increment the version, so no
invalid versions can be published - and prevents unexpected inputs
from causing issues when injected into the "run" steps.