Fix Push to Main Branch Step
Change the "Push to Main and Tag branch step" to more accurately push
to the main and tag branches, then push the new preminor version to
the develop branch.
Also changed the step to run only when the workflow is triggered for
a "feature", because for bug fixes these pushes should be made manually.