Add new React component and clean up files
Add second React component, one for the TreeView that displays
the inherited profiles and catalogs, so the defaultExpanded
prop of TreeView works as expected, allowing us to see the
expanded tree by default and collapse the expanded tree as
desired. TreeView has an expanded prop that doesn't require
a second React component but that doesn't allow for the
collapse of TreeView nodes.

Clean up component, testing, and other files to remove code
that was not necessary and rename functions and variables
to improve readability.