Create a proof of concept for embedded groups
This does a very minimal and incomplete attempt at supporting embedded
groups within a control catalog. As-is, the inner groups render without
much issue; however, there seem to be some assumptions around
expand/collapse as well as how other things get rendered that cause
issues. For example, expanding the outer group expands all inner groups
but collapsing an inner group collapses the "parent" group(s).
Additionally, any of the the bits that are highlighted in yellow (I
forget what they're called) are not rendered properly. Tests need to be
written for this new code as well.