Fix linting errors in YAML files
In general, the issues we had were due to lines being too long or files
missing `---`. at the beginning. I am a big believer in including `---`
so I added that to all files. Additionally, the line length issue seemed
to be a bit rare so I went ahead and just wrapped some lines or
refactored the code, rather than changing the default settings.