Fix errors during loading OSCAL documents
There is an error where an HTTP request that should only occur during
REST mode is enabled. This resulted in an HTTP 404 response (that
actually was perceived as a 200 OK) with an HTML body attempting to be
parsed as JSON. This caused failures up the entire stack. This ensures
we only make that call during REST mode.

Another small error is fixed that only appeared locally. This seems to
be caused by using the spread operator (`...`) on `undefined`.