Add Google Analytics configuration
This adds basic, optional configuration of Google Analytics *if* the
`REACT_APP_GOOGLE_ANALYTICS` environment variable is present at the time
that `npm run build` is executed. This makes it easy to integrate and
fairly self-contained. While the react-ga library's code will be bundled
into the deployed app and the code to do the configuration will be as
well, the initialize won't actually happen unless that value was present
at build time.

At the moment, we just initialize and collect default metrics and
perform a "pageview" action within the App's `useEffect` (which we may
want to consider moving to some sort of an action on the Router but we
don't seem to have a full Router; we only have a Switch).