Add an ErrorBoundary to catch rendering errors
This adds a generic ErrorBoundary which supports catching errors that
weren't previously being caught using the `onError` function/prop
previously. The ErrorBoundary concept, as of React 16, is the preferred
means to set a boundary for handling an error (as it will default to
unmounting the entire app). We, here, just display the generic "Yikes!"
text. Theoretically, we could export `ErrorBoundary`, make it a little
more generic, and use it all across the app (one could imagine rending a
catalog even if we for some reason fail to render a single Back Matter
item).

This is a little different than most our components as we tend to prefer
functional components but the only documentation for an ErrorBoundary
shows it as a class component.