Add concurrency mutex for deployment actions
Our deployments have a tendency to fail if they're started too close to
one another. This is because they'll try to do things like overwrite
each others changes or try to push a version of a package that already
exists. Adding the `concurrency` config attribut ensures that we always
abort the older job if a newer one is starting for the same thing.

We add this at the workflow level because doing it at the `job` level
doesn't result in quite as instantaneous cancellation.