Add support for remote URIs in links
While working on the implementation for #273 it was unclear to me
whether the specification permits diagrams to be referenced directly by
URI or not so I split this out into its own pull request.

The specification specifically states:

> The value of the href can be an internet resource, or a local reference
> using a fragment e.g. #fragment that points to a back-matter resource in
> the same document.
>
> If a local reference using a fragment is used, this will be indicated by
> a fragment "#" followed by an identifier which references an identified
> resource in the document's back-matter or another object that is within
> the scope of the containing OSCAL document.
>
> If an internet resource is used, the href value will be an absolute or
> relative URI pointing to the location of the referenced resource. A
> relative URI will be resolved relative to the location of the document
> containing the link.

This, to me, implied that we should be permitting absolute and relative
links via a URI to internet resources.

To implement this, a "wrapper"-ish function has been created for what
used to be a pure Back Matter lookup function. It handles the logic to
determine whether the link is internal (needs to be looked up in Back
Matter) or an internet resource (is a URI).

There are some incidental changes to code and tests due to the changing
of imports and addition of exported `test-data`.