Persist rest mode toggle across page loads
This makes the configuration of "rest mode" an app-level configuration
that is a parameter to the library components. This matches with how the
Google Analytics is set; additionally, it means that the environment
variable should be set when building the app and that this isn't a
decision made when compiling the library (so it doesn't have to be set
in the CI pipeline!).