Refactor error handling in OSCALLoader
Because `OSCALLoader` itself is the top-level component that has an
`<ErrorBoundary>` as a child, we cannot rely on the `<ErrorBoundary>` to
catch errors further in the stack. Additionally, Error Boundaries do a
poor job of catching asynchronous errors. Instead, we track error state
(which we did prior to #267, though, this implementation is a little
cleaner as we don't try to pass the `onError` further down the call
stack since we _can_ use the `<ErrorBoundary>` for that.

Since we've refactored the logic to handle the error into its own
(reused) function, it's now also trivial to set the `loaded` and
`isResolutionComplete` variables, which means we can also re-enable the
"Reload" button even after an error.

Resolves #283
Resolves #163