Remove GitHub Super-Linter
We've run into issues in the past with the super linter. It evaluates
file one-by-one rather how developers run tests. When run against the
entire codebase it's extremely slow; this causes issues because it means
that configuration changes that may break linting aren't caught (since
rarely do code files get modified at the same time as linting
configuration files). We've had issues with breaking changes in tools
used for linting (such as migrating from jsonlint to eslint).

We lint JSON and JavaScript with `npm run lint`. The only thing we
meaningfully lint outside that is Markdown and YAML. Both of these are
pretty easy to just use the actual tools that Super-Linter is running
anyway.

We might be able to integrate markdownlint into our `npm lint` workflow
but we can't for yamllint (since it's a Python package) so there's not
much value in that as far as I can see.