Ensure npm dependencies can be resolved
We've seen a situation or two now where conflicting package versions
snuck in over time due to incremental `npm install`s and heavy usage of
`npm ci` that very much trusts the `package-lock.json` file. For
example, 8e33c1c4477bd1c3257785e87603e33dbca7b3de started to cause
issues related to this.

This workflow ensures that in a clean environment, without the
`package-lock.json`, an `npm install` will still succeed. No changes are
made to the pull request or to the repository as part of this. It purely
reports success or failure. Using the cache is safe because by default,
only the user npm cache directory is preserved (not `node_modules`).

We don't watch `package-lock.json` because we immediately delete it as
part of this workflow and we only really care about changes to the
version spec/range specified anyway.