Refactor dependency update workflow
Okay so I've made a PR with the same name probably a few times now but
this time, I think I've really done it! This totally refactors the way
we do dependency updates and may (eventually) limit our need to use
Dependabot for _npm_ packages (we'd still need it for other types of
updates). This new workflow runs daily and handles dependency updates in
the root package as well as the example app (so they should no longer
get out of sync).

Notably, this does (for now) require a PAT to open the PR. Eventually,
I'd like to move to using a GitHub App but this feels like a pretty good
starting point.

We also limit React updates to minor versions. Apparently, React 18 is
out and with Storybook and other things, that just sorta wrecks
dependency resolution at the moment. And I am guessing most major
version updates of React will require non-zero effort and need to be
tracked anyway.

As far as Dependabot, I'd like to see these two run side-by-side for a
little bit, with us preferring to merge PRs from this workflow but
seeing how much we "miss" (and how often this workflow fails).