Use a GitHub App instead of a PAT in workflow
A GitHub App is easier for multiple users across the enterprise to
manage _and_ grants access to specific repositories and more finely
scoped permissions than what is possible with a PAT associated with a
user in the organization. We also get to choose a cool name for the app.
This app is named "Easy Dynamics OSCAL Automation" which we can rename
at any time. It currently only has access to this repository; however,
we can trivially grant it access to others.