Add GitHub workflow to automatically assign PRs
Our repositories require that every pull request have two approvals
before they can be merged. Often, what will happen is a team will be
added as a requested reviewer but then the first person who reviews
performs their review "on behalf of" the team. That's pretty cool but it
also means that there are no longer any assigned reviewers! This further
means that for anyone using the [Review requests][1] tab on GitHub, they
don't see that PR still needs to be reviewed.

This new workflow takes care of that by just re-requesting review from
the team until the review threshold has been met. For simplicity in the
implementation, it pauses asking for reviews once someone has requested
changes (but that matches our branch protection rules anyway).

Because assigning team reviewers requires permissions not available on
the `GITHUB_TOKEN`, creating a workflow like this previously would have
been a bit of a complication but we're now able to leverage the same
GitHub App that we're using for our other workflow automation (without
having to request a PAT).

[1]: https://github.com/pulls/review-requested