Pin react-refresh to 0.11.0
When there are conflicting versions of `react-refresh`, `npm run start`
seems to lose its mind (well, not that command directly but something
way deep down in `react-scripts` that's responsible for actually running
the dev server).

For now, let's pin it to 0.11.0 so that people can actually develop and
work. And we can evaluate removing this. The conflicting version that
seems to be coming up at the moment is 0.13.0. We probably could have
picked that version too but this is what most the app was resolving and
so that feels safer.