Fix URL generated for Implemented Requirements API
We were failing to take into consideration the REST URL base so all
calls were relative to `/` which doesn't work. This makes sure the base
is at least considered. There may be a cleaner way to do this that
doesn't require passing `null` around as much but this at least re-uses
the function.