Refactor release workflow
Hopefully this makes it so that this workflow actually works
successfully more option. First, it uses the automation app. Unlike
GitHub Actions, we can authorize this particular app to bypass branch
protection rules and push directly to a branch (`develop` and `main`).
Then we run the workflow as normal, though with slightly more cohesive
steps for setting versions, and push to the relevant branches as
required. We also automatically generate release notes for the release.