Revert changes to `useEffect`
Using two different `useEffects` is necessary because we're really doing
this with two different variables. Because we reference both within the
`useEffect`, they should both be dependencies. But that also means we
end up mutating one of the dependencies during the `useEffect`. So
having these separately actually reduces duplicate/extraneous calls and
isolates the logic.