Perform initial modifications for MUI5
This creates an `AppWrapper` function so that our `ThemeProvider` is the
root of the application and to ensure that `useStyles` is not called
_before_ that. By doing this migration, our `useStyles()` won't get
called until within the `App()` function, which is _after_
`ThemeProvider` is instantiated.

This also includes the `@mui/codemod v5.0.0/preset-safe` codemod on
`src/` and `example/src`. After the codemod ran, a `lint:fix` was also
run to fix styling.

The new MUI5 packages are added but the old MUI4 packages have not yet
been removed.