Refactor after codemods
The codemods made a mess of a few things and introduced `console.error`s
all over the place as well as some broken style. This gets everything as
close to the existing styling as seems practical. A few minor changes to
font sizing and default padding are not worked around but major things
(like the background color) are.