Set app title to "OSCAL Editor" when editing (#398)
* Use the word "Editor" sometimes

* Simplify `useEffect `calls

Co-authored-by: Michael Eisenberg <43248318+mikeisen1@users.noreply.github.com>

* Add missing semicolon

* Revert changes to `useEffect`

Using two different `useEffects` is necessary because we're really doing
this with two different variables. Because we reference both within the
`useEffect`, they should both be dependencies. But that also means we
end up mutating one of the dependencies during the `useEffect`. So
having these separately actually reduces duplicate/extraneous calls and
isolates the logic.

* Cleanup `useEffect` usage with `useMemo`

* Fix lint errors

Co-authored-by: Michael Eisenberg <43248318+mikeisen1@users.noreply.github.com>