Apply URL fixes to top-level profile loads
The current version of the NIST OSCAL content repository still
references the `.xml` files in some cases for `profile.imports.*.href`.
This is likely a bug. For some reason, we were not running into this
before but attempting to load, for example, the rev5 MODERATE baseline
results in a 404 error. This applies the fix not just to further nested
resources but to ones in the top level as well.