Handle the Save on 'Enter' Key
This allows the user to keep the changes in the textbox when they press
the 'Enter' key when editing a textbox.

This feature allows users to save the progress in an editable textfield so
that their changes can be saved. This could be useful if a user wants to
come back an edit a textbox again without having to retype the previous entry.
Pressing 'Enter' will let the user save the current text inside the text box.

This feature is implemented in a similar manner as the saving of the text field
when pressing the save icon to the right of the input box.

I added to the if ladder for the handling the 'Escape' key and added an
onFieldSave to save the changes made by saving the following props:

props.appendToLAst FieldInPath
props.partialRestData
props.editedField
reference.current.value