Reorder app navigation menu
The current ordering is Catalog, SSP, CDef, Profile. Which is an
interesting ordering. Typically, these get listed more inline with the
model of layers of OSCAL, so Catalog, Profile, CDef, SSP.

This is pretty subjective but hopefully sorta prevents time that people
spend searching through the nav bar (not that we have any concrete way
to get metrics for that).

Other possible options would be an alphabetical sorting (Catalog, CDef,
Profile, SSP), or the inverse of this (SSP, CDef, Profile, Catalog)
which also sorta aligns with what we put the most emphasis around
editing and modeling.