OSCALEditableTestField tests for usability (#459)
* OSCALEditableTestField tests for usability

Wrote tests to check that it is not in editing mode after 'Escape' and 'Enter' keys are pressed. Test works for 'Escape' but doesn't work for 'Enter'.

* Test that field is no longer editable

Checks that after both the "Enter" and "Escape" keys are
pressed, the text field is no longer editable. Additionally,
I removed the two tests that had previously been commented
out.