Safely handle undefined fields in profile resolution
Various fields in the OSCAL specification are not required (these are
fields that allow [0, 1] elements). In many cases, we were expecting
those fields to be present. When loading the NIST SP 800-53 rev5
profiles in our viewer, they would fail to render because we required
modify to be present.

This makes some minor changes to profile resolution in order to prevent
similar issues; a more holistic approach to resolution is certainly
still required.