Handle duplicate part ids
The rev5 catalog contained control parts with duplicated ids. Our
strategy for creating keys for these objects was to just use those id's.
This led to duplicated keys warnings in the Viewer.

To deal with that, it was necessary to use the index of each part within
the list to distinguish between parts with the same id. Because using
the index as a key is not usually prefered, extracting the key to a
variable first was necessary.