Make an initial attempt at very basic TypeScript
This integrates the very beginnings of TypeScript into the codebase.
This is far from good. Nearly everything is currently typed as `any` and
the things that aren't may not have their types specified correctly.

This is a proof of concept and we may not actually use this at any point
in time ever.