Make Footer Fixed to Bottom of Viewport
The footer now sticks to the bottom of the viewport at all times.
Furthermore, to accommodate for this change the z-index was adjusted
to a higher level, so the modifications icon doesn't appear on top.
Also, the JSON Editor window is adjusted, so the "SAVE" and "CANCEL"
buttons aren't covered up by the footer but still fill the screen.