Fix component test
After adding markdown to the control implementation descriptions,
a component test in OSCALControlImplementationImplReq.test.js
began failing because it couldn't find the correct link to hover
over. By changing the link name to "a." to hover over, the correct
link is now able to be found and fixes the failing component test.