Fix linting error and update Markdown prose
MarkdownProse implemented MarkupLine. Updated tests to match the
`markupLine` component and implemented the `markupMultiline` part
for testing.

Currently there are 3 tests failing and pushing this commit for
suggestions on how to resolve the tests.

In `testUtil` there is a error since the function is comparing an
undefined value with his text which returns unsuccessfully

I believe the other test is failing due to an improper conversion
from JSX to a string which causes the markdown conversion to go bad
in with the test component.

From what I can see the viewer seems to be working as expected.