Fix testing so after merging in Markup Prose
In oscal Control, removed semicolon between Typography tags so that
extra semicolon does not appear in the editor after each nonplaceholder
text.

In the two test files
Change component testing to test for each of the components using
a regular expression so that it tests that each of those 4 components
can be found on the document instead of the string literal. This also
makes the changes needed to pass the tests after merging in `MarkupProse`

The changes made are equivalent to the previous version of the test.