Remove profile resolution relative URL workarounds (#571)
This cleans up how we handle the creation of URLs. In general, we now make
more effort to cleanup URLs to be absolute and resolved (removing `..` etc)
from the URLs. This results in a more consistent behavior across the
application.

Primarily, this is initiated by usnistgov/oscal-content#60 being resolved;
we had done additional workarounds there to handle incorrect structure in
relative `href`s. Removing the workaround highlighted the possibility for
additional optimizations and improvements.

Closes #505
Closes EasyDynamics/oscal-editor-deployment#128

Co-authored-by: Kyle Laker <kyle@laker.email>
Co-authored-by: Kyle Laker <klaker@easydynamics.com>