Fix hook usage in loader
The loader was using many hooks that resulted in changes every render
due to functions created within the component being used as dependencies
of the `useEffect` hook; though, in some cases those dependencies
weren't in the dependency array! This cleans up the dependency lists and
uses `useCallback` to reduce re-evaluation on every render.