Resolve `act` and `waitFor` callback errors
`fireEvent` does not need to be wrapped in `act` as it is already
wrapped in `act`! React Testing Library has integrated `act` with
its API, including `render` and `fireEvent`.

There is a Testing Library ESLint rule that declares only one
assertion can live in a callback function. It seems if we just move
the multiple `expect` statements outside the `waitFor()` callback,
the same results are achieved and the tests are passed. This will
also allow for any assertions to fail quicker, instead of waiting
for the timeout.