fix: build fails when installing app dependencies
Something went wrong in the `package-lock.json` file for the example
application directory. This is causing unrelated workflows to fail. This
may be a bug in the upgrade-dependencies workflow, but it should not
cause issues with unrelated PRs (and it is). This is also a fix included
in some other dependency-related work that is going to take a minute to
get in, so breaking it out seems prudent.