fix: build fails when installing app dependencies (#650)
Something went wrong in the `package-lock.json` file for the example
application directory. This is causing unrelated workflows to fail. This
may be a bug in the upgrade-dependencies workflow, but it should not
cause issues with unrelated PRs (and it is). This is also a fix included
in some other dependency-related work that is going to take a minute to
get in, so breaking it out seems prudent.

Because the update to `eslint` is now causing build failures for rules
that were always being violated, this adds a bunch of ignores to
`.eslintrc.json` as well. I don't like that but since these were present
before, all we're doing is temporarily continuing to ignore them until
the other open PRs that address them are merged.