Remove ignore rules from ESLint configuration (#660)
In a recent pull request, we disabled many ESLint rules because they
were finally being
detected. This adds corrections for many of them. This is largely a
duplicate of #640;
however, in #640 there are some commits that cause rendering loops and
this PR does not
include those (we will still need to address them but they can be merged
separately).

---------

Co-authored-by: Bradley Fellstrom <bfellstrom@easydynamics.com>