fix(control): change "Guidance" to "Discussion"
"Discussion" is the terminology used in the SP 800-53 PDF document,
whereas "Guidance" is the term used within the OSCAL document. According
to the [800-53rev5 Example][1] from NIST, `guidance` was retained from
an earlier version of the data and tools reading and displaying it are
free to make appropriate adjustments. To keep terminology consistent
between the viewed version of the document and the expected PDF, we
should use "Discussion". Within code, we should continue using
"Guidance" (at least for now) because this is the terminology used in
the source code.

[1]: https://pages.nist.gov/OSCAL/concepts/layer/control/catalog/sp800-53rev5-example/